Refactor content handling in treatment page
Extract the content logic for treatments into a new `ConditionTreatmentContent` component to improve code organization and readability. This refactoring simplifies the `TreatmentForConditionPage` component and enhances maintainability by modularizing the article fetching and rendering logic.

Took 6 minutes
